Only do realistic request errors when we need it
Previously we switched to 'show_exceptions' for all requests in our
tests, which means we return a realistic response instead of raising
an exception. Writing new code with this setting is painful because
it's necessary to debug/inspect the response body and manually look
for an exception/backtrace. This changes 'show_exceptions' back to
false and uses a common 'tag' pattern to enable it where needed.

https://github.com/eliotsykes/rails-testing-toolbox/blob/master/error_responses.rb

This also has the nice side-effect of improving our test performance,
locally and in Docker. I see a consistent improvement in times for
request tests of about 10s.

With: rspec spec/requests  22.40s user 3.16s system 86% cpu 29.413 total

Without: rspec spec/requests  31.52s user 3.26s system 90% cpu 38.490 total