Add file attachment metadata for publications
This adds additional metadata fields for file attachments in order
support the extended workflow for publications. Since these fields
are inherent to the GOV.UK concept of attachments, we think they
should be represented as columns on the model - we also considered
a JSON blob alternative, but it seemed to be mostly disadvantageous.