Add new route to edit featured attachments
Add a new route to edit featured attachments because at the moment the
design for editing inline and featured attachments are quite different.

The new "edit" route will still use the same "update" action as inline
attachments, so a wizard param is being passed in the patch request.
This param will be used to correctly route the user in the next commit.