Create Unpublishing integrity check
This introduce a class to encapsulate all logic for checking that unpublished
editions imported from Whitehall have data that matches what is in Publishing
API. For the first check we expect that the unpublishing type is correct. There
is a difference in wording between the two systems; the state is 'withdrawn' in
Content Publisher and the unpublishing type is 'withdrawal' in Publishing API,
which is why we can't do an equal comparison. Additionally for removed
editions, the Publishing API has 'gone' and 'redirect' unpublishing types,
whereas we don't make this distinction in Content Publisher.