Support unautheticated error viewing
While looking at our previous incident I noticed we had a number of 500
responses for users being served a 404 page. These were occurring when
users were requesting a URL that Rails did not recognise and the
`authenticate_user!` action of warden was raising an exception, trace:

```
UncaughtThrowError
uncaught throw :warden
/data/vhost/content-publisher/shared/bundle/ruby/2.6.0/gems/warden-1.2.8/lib/warden/proxy.rb:134:in `throw'
/data/vhost/content-publisher/shared/bundle/ruby/2.6.0/gems/warden-1.2.8/lib/warden/proxy.rb:134:in `authenticate!'
/data/vhost/content-publisher/shared/bundle/ruby/2.6.0/gems/gds-sso-14.3.0/lib/gds-sso/controller_methods.rb:47:in `authenticate_user!'
/data/vhost/content-publisher/shared/bundle/ruby/2.6.0/gems/activesupport-6.0.2.1/lib/active_support/callbacks.rb:429:in `block in make_lambda
```

This suggests that the Rails exception handler is preventing the throw
mechanism that warden uses to require authentication.

To resolve this I've set up tests for the error actions without
authentication and modified the application layout to handle a user not
existing.