Rename "edit" file attachment to "replace"
In the next story we will be adding a page to add metadata about file attachments.
For images, metadata is added using the "edit" route. To make file attachments more consistent
with images, the "edit" route is being renamed so that it can be used as the metadata path later.