Add featured attachment ordering column
This adds a new column to metadata revisions to store the ordering
of featured attachments. Since we can multiple kinds of attachments
(e.g. HTML attachments), we need a centralised data structure to
refer to all of them, in order to represent them in order.

In planning this work, we considered having a JSON array with items:

  { attachment_type: "FileAttachment", attachment_id: 1 }

As I'm writing the code to use this new column, I realised that:

  - We never use the "attachment_type" value on its own
  - We usually work with the combined UUID e.g. "FileAttachment1"

It turns out we can simplify the way we represent the attachment order
to be an array of the UUIDs we use on the frontend. We only use this
column to order attachments (not to retrieve them, hence the name),
so I can't see any downside to this approach.

Since we expect data in this column to have a defined format, this
also adds a validator for the new column, to help document its contents
and enforce data integrity (in case of errors in our code).