Add a page to record the unique reference
This just displays the page. It doesn't save the data yet.
The "update" page is being added before the "new" because this page will
be linked to directly rather than being added to the middle of flow as
the "new" page will need to be, so is easier to implement.