Apply featured attachment ordering on index page
This adds a new 'featured_attachments' method to a revision that
returns all attachments in the stored order (if one exists). We
then use this on the featured attachments index page, so that the
attachments are listed in the stored order.

While the ordering is primarily a view concern and could go in a
helper, it also seems appropriate for an edition to have a way of
returning all of its featured attachments, so I thought it best to
combine the two concerns.

It's possible the ordering may exclude an attachment e.g. if another
user uploaded a new attachment while the order was being modified. In
this case we do the simplest thing (to_i) to ensure the ordering is
still defined for such attachments; attachments not in the ordering
will (arbitrarily) got at the top of the list.