Add a new feature to reorder featured attachments
This adds a new 'reorder' page/route and the associated interactor
to set a new featured attachment ordering. Since the feature behaves
differently with/out JS, there is a separate scenario for each case.

I did consider splitting these changes across several commits, but
they all seem simple and non-contentious (except maybe the views),
and having them in one commit can make any changes easier to apply.

As with d8b5001c, if an attachment is not found in the ordering
submitted by the user (e.g. if another user uploaded it while the
reordering was going on), we arbitrarily order it at the top.