Add a spec for the reordering interactor
This adds a spec to cover the different kinds of ordering a user may
submit. We've decided to allow any kind of input.

I was unsure how to test the reordering code, as it's in an interactor
and previously interactor tests have covered all/more of their code,
which I don't think is necessary here (it's covered by other tests). So
I've only written tests for the code I think needs covering.