Flag that an updated revison is not in sync
Previously we had repeated tests to check that an updated payload
had been sent when a revision-type change was made e.g. editing content
or tags. Later we decided these tests added little value, since we have
a fallback mechanism to support the user doing this manually if need be,
making the need to call the preview service everywhere 'optional'.

However, the fallback mechanism only works if the edition has flagged
that the revision is not in-sync, using the 'revision_synced' attribute.
This works if we encounter a requirements issue, but not if we just forget
to call the preview service in our code.

This updates the EditDraftEditionService to ensure the edition flags
that a revision is out-of-sync, so that the user sees a fallback
'Preview' button and is not blocked from previewing their changes. I
think this is easier than adding back all the repeated tests.