Add spec to cover an untitled document
Previously we found that a user can abort setting the title for a
document, which results it being shown as '‘’' on many pages, and
as an empty string on the document summary and index pages.

This adds a test to ensure we're using the fallback on the latter
two, as in other scenarios it's unlikely the user is blocked if we
were to forget to use the fallback title.

Since the title of a page is rendered by the layout, we need to render
the application layout explicitly in order to check the title on the
document summary page.