Rewrite I18nCov to generate JSON report
This rewrites a large part of the I18nCov class to separate the
reporting of the coverage from the output (console, file). Using
a machine readable (JSON) data structure for the output file means
we can easily read out the data from it, and will support upcoming
work to enforce 100% I18n coverage.