Pass edition to get file attachment attributes
Previously a 'document' was passed to the helper method we use to
generate the attributes hash/payload for a file attachment. This
changes all uses of the helper to pass an edition, which will make a
further change easier: to make the helper sensitive to the document
type of the edition. We often do 'edition.document' for route helpers,
so this change is broadly consistent with the rest of the app.