Only add extra metadata for featured attachments
Previously we set the unique reference for all attachments, even when
this is not supported by the schema for the document. This updates the
central method we use to generate file attachment attributes, so that
we only include this extra metadata for publications.

Although the unique reference will be nil for other documents (and thus
'compact'ed away), this won't apply to other attributes. It also seems
more robust to explicitly include/exclude these attributes.