Add state problems to integrity checker
We now check the state of the edition in content-publisher versus
publishing-api. The states we're checking against in publishing are not
always a 1 to 1 match as publishing-api doesn't care about the exact
detail of the state in content-publisher. An example of this is an
edition in publishing-api with a state of published could have a state
of published OR published_but_needs_2i in content-publisher. The logic
in expected_state accounts for these cases.