Fix change_history bug
This fixes a bug introduced where by an error is thrown if there is no
change_history for the publishing_api_item, this occurs when the edition
is the first draft. The error is a method error when we call length on
the change_history from within ChangeHistoryCheck.

This commit sets the default change_history to an empty array if it's
not present so length now returns 0 instead of a method error.

The offending line is here -
https://github.com/alphagov/content-publisher/blob/master/lib/whitehall_importer/integrity_checker/change_history_check.rb#L24