Go to new attachment metadata page after upload
This changes the upload behaviour for a featured attachment, so that
we now redirect to the additional page for the user to enter metadata.

Previously the feature test for uploading attachments conflated seeing
the attachment exists (for featured attachments) and the inline previews
for inserting an attachment. This separates the two feature steps,
which makes it easy to test for additional featured attachment metadata.