Don't compare attachment file sizes
We need to exclude the file size of the document from both inline and
linked inline attachments (those with an anchor tag associated) when
comparing the output from Publishing API with that of the proposed Content
Publisher output. The proposed will have a space between the size and
the size type, (i.e. 119 KB) and the output on Publishing API will not
(i.e. 119KB).

First we need to find all possible attachment HTML blocks, their classes
and markup for inline and linked inline attachments on Content Publisher
and Publishing API. From there we can create appropriate tests and
remove them from the comparison process.