Add missing view spec for crop page button
This adds a spec to cover the 'Continue' text of the button on the page
to crop an image, which is meant to reassure the user they don't need to
do anything if their image is already at exact dimensions.

Since 'rendering_context' is set application-wide, it seems reasonable
to define this at the top-level spec helper, alongside 'current_user'.