Skip accessible format notice in integrity checker
There are instances where the proposed payload body for an attachment may include the 'request an accessible format' notice, whereas in the Publishing API body the attachment does not display this notice.

This notice occurs in Whitehall when a checkbox is selected, whereas in Content Publisher, the accessibility notice is determined by the file type.

It may be the case that Whitehall displays this notice, whereas Content Publisher may not. For example, a PDF in Content Publisher will always be deemed in-accessible, whereas in Whitehall, a publisher could mark the content as accessible.

The input is sanitised when passed to `remove_accessible_format_notice` as opposed to during the final comparison in `sufficiently_similar?`. This is because we need to compare the body text 'word content' through removing the html elements or /n characters as the markup differs a lot between the payloads.

Co-authored-by: 1pretz1 <peter.hartshorn@digital.cabinet-office.gov.uk>