Add requirements / validation for ISBN
We allow both ISBN-10 and ISBN-13 variants, but don't want to use
a gem for this validation as the gems we've looked at tend to be
tied to ActiveRecord, or else are not particularly good validators.

We considered using the O'Reilly ISBN regex [1] but this was quite
complex and not restrictive enough (for example, we don't want to
allow a 'ISBN' prefix). Instead we opted for simple regexes taken
from the isbn_validation gem [2], although we've modified them to
remove the '|' character as we don't expect to use that. I took
some example valid and invalid ISBN values from the gem's tests.

ISBN should either be defined and valid, or not defined and not
subject to validation. I have explicitly added a test that checks
that `nil` values are ok.

1: https://www.oreilly.com/library/view/regular-expressions-cookbook/9781449327453/ch04s13.html#I_programlisting4_d1e31888
2: https://github.com/zapnap/isbn_validation