Fix 'Back' link to replace a just-uploaded file
This fixes the 'Back' link on the new metadata page so that going back
and forth between this page and the replace page works correctly in the
context of replacing an uploaded file, while still in the upload workflow.

I've added some missing request tests to cover the (new) behaviour. We
don't have a precedent for testing the 'Back' links specifically, so I'm
going to defer testing that and write 'pain' card about that instead.