Remove 'TODO' to change RSpec/InstanceVariable
Previously we disabled this cop for feature specs with the intention to
explore using 'let' or 'given' instead of instance variables.

This removes the 'TODO', noting instance variables in our feature tests
are assigned *during the test*, so cannot be replaced by 'let' or
'given'. The documentation for this cop also indicates it's meant to be
specific to instance variables assigned in a 'before' block, which could
be replaced with 'let' or 'given'.

describe MyClass do
  before { @foo = [] }
  it { expect(@foo).to be_empty }
end

describe MyClass do
  let(:foo) { [] }
  it { expect(foo).to be_empty }
end

describe MyClass do
  it { <feature_steps> }
  def step_x; ... @foo = input_in_form; ... end
  def step_y; ... expect(result).to include @foo; end
end