Add new modal actions and remove redundant actions
This commit will add new modal actions for confirming
the deletion of images.

It also removes two 'back' actions that were specific to two different
views, there is no need to be so specific so I have replaced them
with a general purpose back action in order to simplify the actions.