Add delete confirmation functionality
This commit adds a view for confirming deletion of an image.
It will add a new method to the image controller and a new
route to enable the confirm delete workflow.

I have also changed the 'confirm-delete' modal action to
camelCase 'confirmDelete' in order to be more consitent
with the rest of the codebase.