Add larger range to compare first published times
We need to add a larger range to compare 'First published' timestamps
due to Whitehall and Content Publisher generating these on the fly
potentially from the first_published_at time. The first_published_at
time is saved on the document in content-publisher and on each edition
in whitehall. The reason these can differ is that whitehall rounds this
time down to the nearest minute after the first edition for each
subsequent edition. This leaves us in a state where we're potentially
comparing the rounded time against the true time and so the range needs
to be larger to accommodate for this.