Import superseded editions with archived history events
While building the status for a number of imported superseded editions, an
error has been raised as there is no event with a "superseded" state in the
edition's revision history. After some investigation into the source data in
Whitehall, it transpires that we have instances where superseded editions have
a final event with the state "archived".  For example:

```
"document_id"=>124314,
"state"=>"superseded",
 ...
"revision_history"=>
  [{"event"=>"create",
    "created_at"=>"2013-03-06T09:37:56.000+00:00",
    "state"=>"imported",
    "whodunnit"=>256},
    ...
   {"event"=>"update",
   "created_at"=>"2013-09-23T13:27:42.000+01:00",
   "state"=>"archived",
   "whodunnit"=>817}],
```

"Archived" looks to be an old edition state which is no longer in use. It was
renamed "Withdrawn" in May 2015:
https://github.com/alphagov/whitehall/commit/95e11e3
However, the documents that we have investigated weren't withdrawn but were
simply superseded by another edition.

The last usage of the "archived" state for edition events was in June 2015, so
we suspect it may be a legacy state, though its unclear why it was being used
instead of 'superseded' in these cases: `Version.where(state:
"archived").maximum(:created_at) => Tue, 02 Jun 2015 16:20:31 BST +01:00`

In order to handle the import of these unusual events for superseded editions,
we have modified the `Create Edition` importer class to look up the last
"archived" event for superseded editions without a "superseded" event.