Support official document type and paper number
This extends the existing metadata page/feature with a new radio
selection for the official document type of the attachment, which
includes a conditional paper number. Requirements checks for the
new data will come in later commits.

We don't have a clear precedent for testing 5 similar happy paths
(one for each radio button). Request specs don't seem appropriate,
since the appearance of the metadata on redirected page is not
related to the request being made. Feature specs aren't appropriate
either, since they are meant to be higher/page-level.

To ensure each of option is covered:

  1. We use an interactor test to ensure the submitted values are
     converted and stored correctly.
  2. We use a unit test to ensure the stored values are presented
     to the Publishing API correctly.
  3. We use a view test to ensure the radio options match up with
     the ones expected by the interactor.
  4. We feature test act paper (numbers) when uploading an ettachment,
     and command paper (numbers) when editing.

Ideally we should have an approach to testing that combines 2 and 3,
and makes 4 arbitrary. But that's out-of-scope for this work.

Another aspect of this commit that's untested is the replay of user
input when the form has requirements issues. This is consistent
with the approach we take in the rest of the app. Manually testing
shows this works, and we've already documented this as something
to address throughout the app, in future work.