Renames 'image_meta' to 'image_with_metadata'
This is to avoid confusion as we plan on creating an 'attachment_meta'
component, which will be literally just attachment metadata (as opposed
to 'attachment with metadata'). 'image_meta' is an image with metadata,
so this renaming is more accurate.

The only place I've not done a rename was the GTM 'save-image-metadata',
which I've kept the same for continuity.