Use 'top_level_description' to render view specs
This replaces the duplication of the partial to render with the
'top_level_description', which couples the partial rendered to the
described text of the view spec. Using 'top_level_description' is
more consistent with our other specs, where we use 'described_class'.

One caveat is that the description of the view specs is no longer
compatible with the direct use of 'render' (they now lack the '_'),
which seems reasonable, since we should be passing locals to partials,
instead of relying on instance variables in the parent view.