Creates 'attachment_meta' component
This has been largely copied from govuk_publishing_components:
https://components.publishing.service.gov.uk/component-guide/attachment

We need the metadata rendered inside the 'Attachment' component,
but don't want the thumbnail or any of the other user-facing
elements such as 'Request an accessible version' or 'Order a copy'
links.

We considered consuming the Attachment component directly and then
hiding the irrelevant portions with CSS, but this is not very
clean or robust and sets a dangerous precedent.

We also considered creating additional parameters to the main
component to hide certain elements, but this quickly becomes
complicated (for example, we also don't want to render a title).

Finally, we considered creating an 'attachment_meta' component in
govuk_publishing_components itself - and doing work to have the
Attachment component consume it rather than duplicate the logic -
but we generally try to avoid cluttering govuk_publishing_components
with additional components until there is an established need for
a component across multiple apps.