Allow official document type to be null
Previously we defaulted the official document type to 'unofficial',
but this is inconsistent with the design/intention to force users to
make an explicit selection. If the user aborts the upload workflow
for an attachment, it's also unclear if a selection was made or not.
This changes the column to default the official doc type to null, so
that 'unofficial' represents an explicit selection by the user.