Remove redundant pre-publish tag check
Previously we duplicated the check that an edition must have a
primary org, which resulted in it appearing twice on the document
summary page: once as a pre-preview issue, and once as a pre-publish
issue. Since the edition must satisfy all pre-preview issues before
publishing, this removes the redundant pre-publish check.