Check for nil change_note when creating edition
When we create a new edition we previously threw a method error if the
current editions change note was nil as we were calling empty? on nil.

As imported editions can have a change note of nil (first edition) we
want to update the conditionals to also cater for nil change notes.