Manage featured attachment ordering
This extends the updater module for file attachments to manage the
featured attachment ordering when file attachments are added or
removed, so that it is always representative of the file attachments
that are currently associated with a revision.

We have arbitrarily decided to append new file attachments to the
existing ordering.