Use a more precise name for ambiguous 'guidance'
Previously we used the term 'guidance_title' to describe a page of
document type selection options, but the term 'title' doesn't relate
to the thing being represented. Unfortunately, we can't use the term
(non-statutory) 'guidance', since this is reserved by an actual document
type. This changes the ID to 'guidance_publications' instead.