Add data integrity validations for JSON columns
This adds a custom validator around the 'contents' and 'tags'
columns, to ensure that we don't store syntactically invalid
data in these otherwise flexible JSON columns. With this change,
we now apply this schema-like approach to all JSON columns.

Note that, although we could use the document type to detect which
tag fields are permitted, this approach doesn't extend to validating
'contents', since the document type isn't accessible from the
associated ContentRevision.