Avoid outputting i18n report unless it is below 100%
Outputting the report is not useful if we already have 100%
coverage, and is momentarily confusing as you need to read
the text to establish whether or not the build succeeded.

By only displaying the report if we fall below 100% coverage,
it is immediately obvious when a local build fails and
requires attention.