Add info about debugging errors with requests
Previously we explored returning to the original testing behaviour,
in a more fine grained way [1], but ultimately it's necessary for
at least some of the tests, and it occasionally confuses people due
to the vague nature of the errors. This should help future developers
understand how to diagnose errors in tests involving a request.

[1]: https://github.com/alphagov/content-publisher/pull/1890