Set all view specs to consistently call template
This applies a consistent approach across all specs where any need to
render a file is done by passing in a template argument. The advantage
of using the template argument as opposed to partial or option omission
is that it will accept the correct path to the template allowing us to
use the top level describe to specify the template in a way that matches
the spec name.