Hide self.class.top_level_description behind described_template
This adds a helper method shared across all view specs of
described_template that pulls in the template name as is described in
the specs description.

I took inspiration from the described_class variable in the name and
chose the word template because, as far as I could tell, this is the
accepted terminology for the underlying file that makes up a view. I
explicitly didn't want to call it view as there is already a method
available of `view` that returns an instance of a view class. I felt
described_view returning a string and view returning an object would be
confusing.

This change comes with some pros and cons, where I believe the pros
outweigh the cons.

Pros:

- opaque, confusing call of `self.class.top_level_description` is hidden
  away from actual specs;
- view tests are easier to read;
- syntax is consistent described_class.

Cons:

- Concept invented in this app that looks like something native in
  RSpec.