Describe templates without mime type and rendering engine
I realised that I based my previous conventions on RSpec Rails 2.2
conventions [1] since that is the ranked higher in Google than the
current version of that documentation.

It transpires that 8 (!) years ago that convention was changed to no
longer be exact paths to the file [2]. This updates all the view tests
to use this convention.

[1]: https://relishapp.com/rspec/rspec-rails/v/2-2/docs/view-specs/view-spec#passing-spec-that-renders-the-described-view-file
[2]: https://github.com/rspec/rspec-rails/commit/ff96af6ba9dbe3a93020dd4e0978235cdad3234b