Switch time helper methods to use Rails time formats
ActiveSupport provides a system to specify named time formatting and has a
convention to do this through a config/initializers/time_formats.rb file
[1]. By using this we can remove the need to have helper functions to
provide the time and rely on passing a symbol to the `to_s` method of a
time. This change was inspired by https://andycroll.com/ruby/using-strftime-in-a-rails-view-is-probably-a-mistake/
that I'd been meaning to try out and finally found myself working on a
rather quiet Saturday shift.

As always with something like this, a painful part of this was deciding what to
name the formatters. I felt it was best to go with very similar to
naming as what we'd had with the helper. This has a downside that it
clobbers the existing formatting of `Time::DATE_FORMATS[:time]` but
since we make no use of that it has no affect on the application.

I also resolved some trailing comma issues in hashes so they were
consistent with our linting rules (which aren't enforced on erb files)

[1]: https://github.com/rails/rails/blob/b2eb1d1c55a59fee1e6c4cba7030d8ceb524267c/activesupport/lib/active_support/core_ext/time/conversions.rb#L43-L50