Fix allowing empty primary orgs
Previously we added validations on the JSON stored for the 'tags',
which has uncovered a bug where '[""]' was being stored when the
user had no primary publishing organisation (or cleared it). This
fixes the field to properly cleanup the input and also changes it
to behave similarly to a MultiTagField when the param isn't there.