Fix replay of submitted data with issues
Previously this was using the non-existant 'tags' params to check
for values to replay. Since the value to replay could be nil/blank,
this uses the 'key?' method of params to correctly replay any param
that was submitted, even if it was invalid/blank.

Related pain card: https://trello.com/c/hccZ9cuf/1715-we-dont-test-that-invalid-inputs-are-highlighted-values-retained