Fix Sentry should capture block
It transpires that this block is called not just with an exception
class in the report of an error it is also called with a Raven::Event
which does not respond to a cause event.

This updates the error filtering to be resilient to the object being
passed in to not conform to a particular interface.

Expected type inputs to this lambda would be a Ruby Exception, a
Raven::Event and a String.