Fix missing link for image alt text issue
This also adds a test for the 'requirements' partial. Although there
are lots of potential links to fix requirements, most of them are
shown elsewhere on the summary page, so are arguably for convenience
only. Previously, we found that an alt text issue was blocking for
users, without a link to the page to fix it.

This also removes redundant linking capability for 'credit' and
'caption', since issues for these fields will never appear on the
document summary.

Bug introduced in: 24dce0940cb7b1195a22bfe5226bb89872f5c313