Use a common interface for requirements checkers
Previously we had a range of interfaces for checkers: some received
their arguments in on 'initialize', some on '_issues'; some had
multiple methods to check for issues, some just one. This introduces
a common interface for requirements checkers: each checker has a
single '.call' method with arguments for the checker; any checkers
with multiple methods have been split into separate classes.

Previously the checker classes were in a flat namespace, with some
checkers being used in multiple contexts: on form submission, before
preview, and before publish. It was hard to understand which checker
was used for what. This builds on the separation of methods
into classes by arranging them in a namespace that reflects the context
in which the checker is used: 'form', 'preview' or 'publish'.

Some more minor changes:

  - The check for a blank file attachment is now done in the interactor
    that calls the checker, so the checker can also be used for when
    the title (and not the file, on the same form) is updated.

  - The flag to 'rescue_api_errors' has been lifted to the top-level
    checker that receives it, which then enables it to iterate over
    the lower-level checkers, instead of calling each one specifically.
    While it's not ideal for the exception to be displaced from its
    origin, we could easily address this in future by introducing a
    proxy exception, such as 'Requirements::CheckerUnavailable'.