Remove repetition of %{time} on %{date}
We use the same date time formatting convention throughout Content
Publisher, which is typically "%{time} on %{date}". We've slipped into
an inconsistency where we have an available formatting that shows a time
value as a time on date value, but only used it for a few times across
the app.

This commit changes this so that we use the time_on_date formatting for
all the places where this convention is used. This typically reduces the
number of arguments needed to load a translation by one. I chose to use
the variable name `datetime` to represent when we were using a date and a
time together, this was chosen because we had precedence for this name
in the app and because it didn't specify the formatting. I decided not
to use `time` even though it was semantically appropriate as I felt the
lack of date in it's name left it's purpose confused.

I had hoped that this might mean I could remove the `time` formatter but
it's still needed in one location: the form to set a scheduling time.