Replace special time_at_date format rule
We only have usage of the time at date formatting on the timeline entry
screen. Since this is only used on the timeline entry page it didn't
seem to be necessary to be set-up as a time formatting rule. This also
allowed getting rid of the somewhat strange translation key of "by".

I'm not 100% sure this is a great idea, as a couple of design decisions
may bring this formatter rule right back, but it does at least
discourage initial spread of it while we have a clear preference.

It did feel a bit of a bummer repeating this logic in both templates,
however I recall there being an idea to replace this with a component
which would be a better way to remove the duplication.