DRY-up issues in requirements checkers
Previously each checker had to instantiate a CheckerIssues container,
and return it at the end of each method. This moves that responsibility
to the new base class, so that the checkers can focus on populating
the CheckerIssues container. Note that some checkers need to set the
issues, when they aggregate them from lower-level checkers.