Update previews when changing schedule proposal
Since 9392d8dcd5a0524bfd2c21c8e8c485b57cd84cb3 we now mark all editions
with revision_synced of false when using the EditDraftEditionService.
This means that when we change proposed publish time we are shown the
actions view to indicate a preview was not successful.

Previously I believe that we didn't update the preview on these actions
as these changes didn't actually affect the Publishing API. However it's
probably best not to treat this as something special.