Restrict requirements to featured attachments
Previously we added new requirements checks that should only apply
to featured attachments, as defined by the type of document they're
associated with. Although the checks primarily run in the context
of the form, which is only available for featured attachments, we
also run these checks at the 'edition level' before publish. This
fixes the edition level check to reflect the type of attachment.