Use 'push' to add checker issues together
Previously we used '+=' to add multiple sets of issues together. While
the '+' operator used to be a good fit, it's now surprising to see a
checker reset the 'issues' instance variable declared by the new base
Checker class. This replaces the use of '+=' with 'push *'.

The choice of 'push *' is analogous to the way multiple elements can be
added to an Array (in-place). Other variations we considered:

- 'merge', which I find surprising because we're not de-dupping
  anything, and the operation isn't immutable (like it is for Hash)

- 'push' (no splat), which I find surprising because it's ambiguous if
  one should pass an Issue, or a CheckerIssues.