Make header and footer links consistent with ADR-4
In [ADR-4][] it was defined that link text and labels would not be
defined as configurable locale copy. This changes these header and
footer links to be consistent with this and resolves an inconsistency
where the feature test (spec/features/publisher_information/visit_publisher_information_pages_spec.rb)
had exact copy used rather than using locale lookups.

As part of this I intentionally renamed "GOV.UK Status" to "GOV.UK
status" since "GOV.UK Status" does not seem to be the name of a specific
product.

[ADR-4]: https://github.com/alphagov/content-publisher/blob/master/docs/adr/0004-editing-microcopy.md