Explicitly specify which gems can update
Previously we specified versions of each dependency in the Gemfile,
and relied on DependaBot to automatically update them. This changes
the specified versions to explicitly state which dependencies can
freely update, and which remain relatively static.

We are proposing to adopt a new usage of DependaBot, which will not
raise a PR to update a dependency if that would require a change in
the Gemfile. The aim of this change is to limit the time we spend
updating dependencies, by only updating the following:

- Our internal dependencies
- Rails
- PostGreSQL
- RSpec

We can always update any dependency manually, at the point of need.
The reason for auto-updating some dependencies are:

- We write a lot of code for/with them. Postponing updates could
make it impractical to apply those updates in the future.

- We frequently benefit from the using the latest releases of these
dependencies.

- We frequently refer to the documentation for these dependencies,
which is usually limited or defaulted to the latest version.